requirements_py311.txt
# Requirements for Python 3.11 (Windows)
alpaca-trade-api==3.1.1
pandas==2.2.2
numpy==1.26.4
ta==0.10.2
python-dotenv==1.0.1
pytz==2024.1
websockets>=9.0,<11
# Note: aiohttp is resolved automatically by alpaca-trade-api to 3.8.2 on py311
